Started the screen flow
created the general format of the screens. Added 3 buttons to the
mainActivity; Login, Create Account; Login as guest. Login as guest
takes you to a blank activity which will hold all the feed buttons
because a guest can view the feeds. The login and Create account buttons
will get sent to login activities. The create account activity will have
to modified to to the point where it is Create Account but it is close.
The formatting for these buttons still have to be worked on. I am also
in the process of adding a home button to each of the activities that
will take you back to the mainActivity.