Created basic screen flow
I have finished creating the screen flow for the guest, right now the
Feed Me and Tell Me activities just have text on them explaining what
they are and a home button that takes you back to Guest Main. I have put
code to when clicking sign in or register on create Account to send to
Main screen and it wont go, so i will look into this soon.